agregacion de nav superior, ruta al detail de personas y crud de personas por la mitad

+++ backend/src/main/java/unpsjb/labprog/backend/model/Persona.java
@@ -15,7 +15,7 @@ import lombok.Setter;
 @NoArgsConstructor
 public class Persona {
     @Id
-    private Long dni;
+    private int dni;
 
     private String cuil;
     private String nombre;
